### 2025-05-28 #8299 @editorialbot
[REVIEW]: HINA: A Learning Analytics Tool for Heterogenous Interaction Network Analysis in Python
**Submitting author:** @shihuife (<a href="http://orcid.org/0000-0002-5572-276X">shihui Feng</a>)
**Repository:** https://github.com/SHF-NAILResearchGroup/HINA
**Branch with paper.md** (empty if default branch):
**Version:** v1.4
