### 2025-05-31 SHF-NAILResearchGroup/HINA#5 @etShaw-zh
[JOSS] Insufficient Test Coverage
Very useful and interesting package.

While reviewing the codebase, I identified a typo in a variable name which has now been corrected. ref: #4
After running the full test suite using pytest, all tests passed successfully. However, the current test coverage remains insufficient.

![Image](https://github.com/user-attachments/assets/3dcd15d1-9927-44a7-804e-3c07137ed7b0)

Action Suggested:

- Increase unit test coverage, especially for less-tested modules.
- Consider adding edge case tests for robustness.
- Optionally, suppress or resolve warnings for a cleaner test output.

Recommend integrating [Codecov](https://about.codecov.io/) to help reviewers easily monitor and assess test coverage over time.

ref: https://github.com/openjournals/joss-reviews/issues/8299
