### 2025-05-31 #8299 @etShaw-zh
Hi, @juliaferraioli @shihuife

I have completed my initial thorough review of both the `HINA` codebase and the accompanying manuscript, focusing primarily on documentation improvements. Relevant contributions include:

Issue:

- [Insufficient Test Coverage](https://github.com/SHF-NAILResearchGroup/HINA/issues/5)

PRs:

- [Fix typo diversity](https://github.com/SHF-NAILResearchGroup/HINA/pull/4)
- [Paper Documentation Improvements](https://github.com/SHF-NAILResearchGroup/HINA/pull/2#issue-3049463668)

A critical blocker for JOSS acceptance is insufficient test coverage (see SHF-NAILResearchGroup/HINA#5). Once the test coverage concerns are resolved, I would be comfortable checking off those corresponding review items.
