### 2025-06-09 SHF-NAILResearchGroup/HINA#5 @shihuife
@etShaw-zh Thank you for the thorough review and suggestions. We have addressed all suggested actions and integrated Codecov.
