### 2025-06-09 SHF-NAILResearchGroup/HINA#5 @etShaw-zh
Thank you for your contribution. I feel that my concerns have been addressed.

[![Coverage Status](https://codecov.io/gh/SHF-NAILResearchGroup/HINA/branch/main/graph/badge.svg?token=GLAVYYCD9L)](https://codecov.io/gh/SHF-NAILResearchGroup/HINA)

ref: https://github.com/openjournals/joss-reviews/issues/8299#issuecomment-2923924655
