### 2025-06-09 #8299 @etShaw-zh
@shihuife Thank you for your contribution. This is a valuable tool for learning analytics.  👍👍
@juliaferraioli The author has resolved my concerns. I recommend this contribution for publication in JOSS if other reviewers have no issues.
[![Coverage Status](https://codecov.io/gh/SHF-NAILResearchGroup/HINA/branch/main/graph/badge.svg?token=GLAVYYCD9L)](https://codecov.io/gh/SHF-NAILResearchGroup/HINA)
