### 2025-06-23 SHF-NAILResearchGroup/HINA#7 @drj11
Wrong import for mesoscale Demo
(i am reviewing for https://github.com/openjournals/joss-reviews/issues/8299 )

The example code for the `mesoscale` module imports the wrong thing https://hina.readthedocs.io/en/latest/Modules/mesoscale.html#example-code

    from hina.dyad import prune_edges

which means that the rest of the demo does not run:

```
>>> result_no_fix = hina_communities(B)
Traceback (most recent call last):
  File "<python-input-28>", line 1, in <module>
    result_no_fix = hina_communities(B)
                    ^^^^^^^^^^^^^^^^
NameError: name 'hina_communities' is not defined
```
