### 2025-06-24 #8299 @drj11
this is a comment to the authors and not directly relevant to the matter of reviewing, but i wonder if the authors have come across HyperNetX https://joss.theoj.org/papers/10.21105/joss.06016 (which i only know about because i review it for JOSS!); it allows modelling of the _hypergraph_, that is, graphs on the _sets of nodes_; for example in this case, between the identified sets of learners.
