### 2025-06-24 #8299 @shihuife
Hi @drj11, many thanks for taking the time to review our package!

Thanks a lot for pointing us to HyperNetX. The uniqueness of HINA is that we provide tailored network-based solutions for analyzing learning process data. The individual, dyadic, mesoscale methods are specifically designed for modelling individual- and group- learning processes, distinguishing among different categories of nodes simultaneously by utilising the HIN representation. This is required for our diversity measure and nonparametric clustering method, and simplifies visualization in many cases for the student interaction setting.

In our paper, we compared our package with general network analysis packages and network inference packages, and have now added references to HyperNetX and other hypergraph libraries (XGI, HGX). Many thanks!
