### 2025-07-09 SHF-NAILResearchGroup/HINA#10 @drj11
Running tests with `pytest` requires `httpx`
```
(venv) ; which pytest
/Users/drj/repo/HINA/venv/bin/pytest
(venv) ; pytest hina/
platform darwin -- Python 3.13.3, pytest-8.4.1, pluggy-1.6.0
rootdir: /Users/drj/repo/HINA
plugins: anyio-4.9.0, dash-3.0.0
collected 35 items / 1 error

==================================== ERRORS ====================================
_______________________ ERROR collecting hina/app/tests ________________________
venv/lib/python3.13/site-packages/starlette/testclient.py:30: in <module>
    import httpx
E   ModuleNotFoundError: No module named 'httpx'

/usr/local/Cellar/python@3.13/3.13.3/Frameworks/Python.framework/Versions/3.13/lib/python3.13/importlib/__init__.py:88: in import_module
           ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
<frozen importlib._bootstrap>:1387: in _gcd_import
    ???
<frozen importlib._bootstrap>:1360: in _find_and_load
    ???
<frozen importlib._bootstrap>:1331: in _find_and_load_unlocked
    ???
<frozen importlib._bootstrap>:935: in _load_unlocked
    ???
venv/lib/python3.13/site-packages/_pytest/assertion/rewrite.py:186: in exec_module
    exec(co, module.__dict__)
hina/app/tests/conftest.py:7: in <module>
    from fastapi.testclient import TestClient
venv/lib/python3.13/site-packages/fastapi/testclient.py:1: in <module>
    from starlette.testclient import TestClient as TestClient  # noqa
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
venv/lib/python3.13/site-packages/starlette/testclient.py:32: in <module>
    raise RuntimeError(
E   RuntimeError: The starlette.testclient module requires the httpx package to be installed.
E   You can install this with:
E       $ pip install httpx
=========================== short test summary info ============================
ERROR hina/app/tests - RuntimeError: The starlette.testclient module requires the httpx package to...
!!!!!!!!!!!!!!!!!!!! Interrupted: 1 error during collection !!!!!!!!!!!!!!!!!!!!
============================== 1 error in 21.46s ===============================
```

With that installed, after `pip install httpx`, the tests appear to run to completion:

```
(venv) ; pytest hina/
platform darwin -- Python 3.13.3, pytest-8.4.1, pluggy-1.6.0
rootdir: /Users/drj/repo/HINA
plugins: anyio-4.9.0, dash-3.0.0
collected 47 items

hina/app/tests/test_api.py .....                                         [ 10%]
hina/app/tests/test_utils.py .......                                     [ 25%]
hina/construction/tests/test_network_construct.py ....                   [ 34%]
hina/dyad/tests/test_significant_edges.py ..........                     [ 55%]
hina/individual/tests/test_diversity.py ...                              [ 61%]
hina/individual/tests/test_quantity.py ..                                [ 65%]
hina/mesoscale/tests/test_clustering.py .....                            [ 76%]
hina/utils/tests/test_graph_tools.py ....                                [ 85%]
hina/visualization/tests/test_network_visualization.py .......           [100%]

============================== 47 passed in 4.18s ==============================
```

For the record, my venv installed `httpx-0.28.1`

For my JOSS review, can you please update the documentation.
