### 2025-07-09 SHF-NAILResearchGroup/HINA#10 @shihuife
Many thanks for the suggestion! We have added the httpx installation requirement to the document.
