### 2025-07-10 a5a8cd0
Update paper.md
