### 2025-07-10 #8299 @drj11
(For my part) I have completed my checklist and my review. There are no further comments that i require to be addressed. I recommend that this paper by accepted by JOSS for publication.

Great job!

@juliaferraioli
