### 2025-07-15 972717d
Create CITATION.cff
