### 2025-07-15 9589138
Update CITATION.cff
