### 2025-07-15 feed814
Update CITATION.cff
