### 2025-07-15 #8299 @shihuife
@juliaferraioli   Thank you very much for handling this review! We have released the software and the version tag of the archived version is: SHF-NAILResearchGroup/HINA-v1.4, Zenodo DOI: 10.5281/zenodo.15940278.
