### 2025-07-23 dbcb0ce
adjust paper.bib
