+++ Paper/paper.bib
@@ -49,7 +49,7 @@
 
 @article{landry2023, 
   author = {Nicholas W. Landry and Maxime Lucas and Iacopo Iacopini and Giovanni Petri and Alice Schwarze and Alice Patania and Leo Torres},
-  title = {XGI: A Python package for higher-order interaction networks}, journal = {Journal of Open Source Software},
+  title = {XGI: A {P}ython package for higher-order interaction networks}, journal = {Journal of Open Source Software},
   doi = {10.21105/joss.05162}, 
   url = {https://doi.org/10.21105/joss.05162}, 
   year = {2023}, 
@@ -83,7 +83,7 @@
 }
 
 @article{peixoto2014graphtool,
-   title = {The graph-tool python library},
+   title = {The graph-tool {P}ython library},
    url = {http://figshare.com/articles/graph_tool/1164194},
    doi = {10.6084/m9.figshare.1164194},
    urldate = {2014-09-10},
@@ -102,7 +102,7 @@
 
 @article{praggastis2024,  
   author = {Brenda Praggastis and Sinan Aksoy and Dustin Arendt and Mark Bonicillo and Cliff Joslyn and Emilie Purvine and Madelyn Shapiro and Ji Young Yun}, 
-  title = {HyperNetX: A Python package for modeling complex network data as hypergraphs}, 
+  title = {HyperNetX: A {P}ython package for modeling complex network data as hypergraphs}, 
   journal = {Journal of Open Source Software}, 
   year = {2024}, 
   publisher = {The Open Journal}, 
