### 2025-07-23 190b9a0
Update paper.md
