+++ Paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: '**HINA: A Learning Analytics Tool for Heterogenous Interaction Network Analysis in Python**'
+title: 'HINA: A Learning Analytics Tool for Heterogenous Interaction Network Analysis in Python'
 
 tags:
   - Python
