### 2025-07-23 #8299 @crvernon
🥳 Congratulations on your new publication @shihuife! Many thanks to @juliaferraioli for editing and @etShaw-zh, @rtbs-dev, and @drj11 for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
